refactor(find): consolidate pure detectors into find.py
Move find_fight_mode_icon and find_post_battle_button from nav.py, and detect_upgradable_cards from upgrade_state.py, into a new leaf-style pyclashbot/bot/find.py. find.py imports only from pyclashbot.detection.*, pyclashbot.utils.*, pyclashbot.bot.coords, and pyclashbot.bot.state_detect (for the pixel_indicates_upgradable predicate, documented as an allowed exception since state_detect is itself leaf-style and does not import from find).
