+++ pyclashbot/bot/upgrade_state.py
@@ -14,26 +14,16 @@ from pyclashbot.bot.coords import (
     UPGRADE_RETURN_TO_MAIN_COORD_1,
     UPGRADE_RETURN_TO_MAIN_COORD_2,
 )
+from pyclashbot.bot.find import detect_upgradable_cards
 from pyclashbot.bot.nav import (
     get_to_card_page_from_clash_main,
     select_mode,
     wait_for_clash_main_menu,
 )
-from pyclashbot.bot.state_detect import check_if_on_clash_main_menu, pixel_indicates_upgradable
+from pyclashbot.bot.state_detect import check_if_on_clash_main_menu
 from pyclashbot.detection.image_rec import pixel_is_equal
 from pyclashbot.utils.cancellation import interruptible_sleep
 from pyclashbot.utils.logger import Logger
-
-
-def detect_upgradable_cards(emulator):
-    img = emulator.screenshot()
-    upgradable = []  # it will be garbage collected right?
-
-    for i, (x, y) in enumerate(UPGRADE_POINTS, start=1):
-        if pixel_indicates_upgradable(img[y][x]):
-            upgradable.append(i)
-
-    return upgradable
 
 
 def upgrade_card(emulator, upgradable, logger: Logger):
